// frontend/src/api/generated/models/issueLinkRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { LinkedIssue } from './linkedIssue';

export interface IssueLinkRead {
  id: number;
  relation: string;
  issue: LinkedIssue;
  created_at: string;
}